+++ web/src/handlers/index.ts
@@ -68,7 +68,7 @@ export const order = {
     placeOrder: placeOrder,
     createTransaction: createTransaction,
     getMyOrders: getOrders,
-    gerOrderByID: getByOrderID,
+    getOrderByID: getByOrderID,
     getInvoiceURI: getInvoiceURI,
     isOrderComplete: isOrderComplete,
     getShippingInformation: getShippingInformation,
