+++ web/src/pages/store/orders/index.tsx
@@ -14,63 +14,12 @@ import useGetOrders from "@hooks/orders/use-get-orders"
 import RenderIcon from "@components/render-icon"
 import icons from "@/icons"
 import IconButton from "@mui/material/IconButton"
-import { useCallback, useEffect, useMemo, useState } from "react"
+import { useMemo, useState } from "react"
 import { Box, Button, Chip, Popover, Stack, Typography } from "@mui/material"
 import { order } from "@api"
 import useGetStatusColor from "@hooks/shipping/use-get-status-color"
-import useConnectPaddle from "@hooks/orders/use-connect-paddle"
-import useGetCustomerID from "@hooks/users/use-get-customer-id"
 import Render from "@components/render"
-
-const useCompleteTransaction = (cb: () => void) => {
-    const customerID = useGetCustomerID()
-    const { paddle, event } = useConnectPaddle(customerID)
-
-    const completeOrder = useCallback(
-        async (txnID: string) => {
-            try {
-                if (paddle) {
-                    paddle.Checkout.open({
-                        settings: {
-                            displayMode: "overlay",
-                            variant: "multi-page"
-                        },
-                        transactionId: txnID
-                    })
-                }
-            } catch (error) {
-                const err = error as Error
-                console.log(err)
-            }
-        },
-        [customerID, paddle]
-    )
-
-    const { name, data } = event || {}
-    const { transaction_id } = data || {}
-
-    useEffect(() => {
-        if (name === "checkout.completed") {
-            const checkStatus = async (txnId: string) => {
-                const isComplete = await order.isOrderComplete(txnId)
-                if (isComplete) {
-                    cb()
-                    paddle?.Checkout.close()
-                }
-            }
-
-            if (!transaction_id) return
-
-            const intervalID = setInterval(() => {
-                checkStatus(transaction_id)
-            }, 3000)
-            return () => {
-                clearInterval(intervalID)
-            }
-        }
-    }, [name, transaction_id, paddle])
-    return completeOrder
-}
+import useCompleteTransaction from "@hooks/orders/use-complete-transaction"
 
 const Orders = () => {
     const { data: orders, refetch } = useGetOrders()
@@ -146,11 +95,13 @@ const Orders = () => {
                                                 }
                                                 show={
                                                     <IconButton
-                                                        onClick={() =>
-                                                            order.gerOrderByID(
-                                                                row.transaction_id
-                                                            )
-                                                        }
+                                                        onClick={async () => {
+                                                            const data =
+                                                                await order.getOrderByID(
+                                                                    row.transaction_id
+                                                                )
+                                                            window.open(data)
+                                                        }}
                                                     >
                                                         <RenderIcon
                                                             icon={
